requirements.txt
flask>=2.0.0
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
python-chess>=1.9.0
cairosvg>=2.7.0
tensorflow>=2.7.0
pytest>=6.0.0
